Update NavBar
This is code that Lucas used for the previous project that wasn't merged
