+++ frontend/package.json
@@ -61,7 +61,7 @@
     "eslint-plugin-react-refresh": "^0.3.4",
     "nodemon": "^2.0.6",
     "postcss": "^8.4.23",
-    "react-router-dom": "^6.11.2",
+    "react-router-dom": "^6.16.0",
     "tailwindcss": "^3.3.2",
     "typescript": "^5.0.2",
     "vite": "^4.3.2"
