+++ frontend/src/components/auth/SignUp/SignUpForm.tsx
@@ -47,8 +47,6 @@ const SignUpForm: React.FC<SignUpFormProps> = () => {
     setRepeatPassword(e.target.value);
   };
 
-  //   if (user) return <Redirect to="/" />;
-
   return (
     <form className="main-user-signup" onSubmit={onSignUp}>
       <div className="errors-list-signup">
@@ -112,7 +110,6 @@ const SignUpForm: React.FC<SignUpFormProps> = () => {
         <button className="user-buttons" type="submit">
           Sign Up
         </button>
-        {/* <Demo /> */}
       </div>
     </form>
   );
